Modified DatabaseHelper to use a new database for every test.

+++ src/test/java/net/maltera/daranable/edinet/DatabaseHelper.java
@@ -6,32 +6,34 @@ import java.util.Random;
 
 public class DatabaseHelper {
 	private final Properties vanillaProps;
-	private final Properties standardProps;
+	private final Random rand = new Random();
 	
-	private final File testDir;
+	private File testDir;
 	
 	public DatabaseHelper() {
-		Random rand = new Random();
+		vanillaProps = System.getProperties();
+	}
+	
+	public void prepare() {
 		testDir = new File( System.getProperty( "user.dir" ), "edinet_test_" 
 				+ String.format( "%010d", rand.nextInt( Integer.MAX_VALUE ) ) );
 		testDir.mkdirs();
 		
-		vanillaProps = System.getProperties();
+		Properties props = (Properties) vanillaProps.clone();
+		props.setProperty( "os.name", "Testing" );
+		props.setProperty( "user.home", testDir.getAbsolutePath() );
 		
-		standardProps = (Properties) vanillaProps.clone();
-		standardProps.setProperty( "os.name", "Testing" );
-		standardProps.setProperty( "user.home", testDir.getAbsolutePath() );
-	}
-	
-	public void prepare() {
-		System.setProperties( standardProps );
+		System.setProperties( props );
 	}
 	
 	private static void deleteRecursive (File target) {
-		for (File file : target.listFiles()) {
-			if (file.isDirectory())
-				deleteRecursive( file );
-			else file.delete();
+		File[] contents = target.listFiles();
+		if (null != contents) {
+			for (File file : contents) {
+				if (file.isDirectory())
+					deleteRecursive( file );
+				else file.delete();
+			}
 		}
 		
 		target.delete();
